requirements.txt

tensorflow>=2.9      # If using GPU, install tensorflow-gpu matching your CUDA
numpy
pandas
matplotlib
scikit-learn
Pillow
opencv-python
